+++ test/naturalBreaksManager.js
@@ -16,6 +16,7 @@ describe('naturalBreaksManager', function () {
 
   it('should be running with default settings', () => {
     naturalBreaksManager.isOnNaturalBreak.should.be.equal(false)
+    naturalBreaksManager.isSchedulerCleared.should.be.equal(false)
     naturalBreaksManager.usingNaturalBreaks.should.be.equal(true)
   })
 
@@ -25,6 +26,7 @@ describe('naturalBreaksManager', function () {
     naturalBreaksManager = null
     naturalBreaksManager = new NaturalBreaksManager(settings)
     naturalBreaksManager.isOnNaturalBreak.should.be.equal(false)
+    naturalBreaksManager.isSchedulerCleared.should.be.equal(false)
     naturalBreaksManager.usingNaturalBreaks.should.be.equal(false)
   })
 
@@ -34,6 +36,7 @@ describe('naturalBreaksManager', function () {
     naturalBreaksManager = null
     naturalBreaksManager = new NaturalBreaksManager(settings)
     naturalBreaksManager.isOnNaturalBreak.should.be.equal(false)
+    naturalBreaksManager.isSchedulerCleared.should.be.equal(false)
     naturalBreaksManager.usingNaturalBreaks.should.be.equal(true)
   })
 
@@ -41,12 +44,14 @@ describe('naturalBreaksManager', function () {
     naturalBreaksManager.stop()
     naturalBreaksManager.start()
     naturalBreaksManager.isOnNaturalBreak.should.be.equal(false)
+    naturalBreaksManager.isSchedulerCleared.should.be.equal(false)
     naturalBreaksManager.usingNaturalBreaks.should.be.equal(true)
   })
 
   it('should stop when stop()', () => {
     naturalBreaksManager.stop()
     naturalBreaksManager.usingNaturalBreaks.should.be.equal(false)
+    naturalBreaksManager.isSchedulerCleared.should.be.equal(false)
     naturalBreaksManager.isOnNaturalBreak.should.be.equal(false)
     naturalBreaksManager.idleTime.should.be.equal(0)
   })