+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ## Fixed
-- multi-line status messages in Tray (Mac, Windows) 
+- multi-line status messages in Tray (Mac, Windows)
+
+## Added
+- show Paused icon when Breaks are paused in Natural break
 
 ## [0.99.2] - 2020-04-27
 ### Removed